delete codegenOpts, fully respects config.config
Use @apollo/client@3, @graphql-codegen/typescript-react-apollo@2
