+++ package.json
@@ -36,6 +36,7 @@
     "yaml": "^1.7.2"
   },
   "devDependencies": {
+    "@apollo/client": "^3.1.3",
     "@apollo/react-common": "^3.1.3",
     "@apollo/react-components": "^3.1.3",
     "@apollo/react-hooks": "^3.1.3",
@@ -48,7 +49,7 @@
     "@graphql-codegen/plugin-helpers": "^1.17.7",
     "@graphql-codegen/typescript": "^1.17.7",
     "@graphql-codegen/typescript-operations": "^1.17.7",
-    "@graphql-codegen/typescript-react-apollo": "^1.17.7",
+    "@graphql-codegen/typescript-react-apollo": "^2.0.5",
     "@graphql-codegen/typescript-resolvers": "^1.12.2",
     "@graphql-tools/graphql-file-loader": "^6.0.16",
     "@graphql-tools/import": "^6.0.16",
@@ -76,10 +77,6 @@
     "@typescript-eslint/eslint-plugin": "^2.20.0",
     "@typescript-eslint/parser": "^2.20.0",
     "@typescript/vfs": "^1.0.1",
-    "apollo-cache": "^1.3.4",
-    "apollo-client": "^2.6.8",
-    "apollo-link": "^1.2.13",
-    "apollo-utilities": "^1.3.3",
     "babel-jest": "^26.0.1",
     "babel-loader": "^8.0.6",
     "babel-preset-env": "^1.7.0",
