+++ src/lib/exec-context.ts
@@ -1,12 +1,10 @@
 import { ConfigTypes } from './config';
-import createCodegenOpts, { PartialCodegenOpts } from './create-codegen-opts';
 import { getCacheFullDir } from './paths';
 
 export type ExecContext = {
   cwd: string;
   config: ConfigTypes;
   configHash: string;
-  codegenOpts: PartialCodegenOpts;
   cacheFullDir: string;
 };
 
@@ -15,12 +13,10 @@ export default function createExecContext(
   config: ConfigTypes,
   configHash: string,
 ) {
-  const codegenOpts = createCodegenOpts(config);
   const cacheFullDir = getCacheFullDir(cwd, config.cacheDir);
   return {
     cwd,
     config,
-    codegenOpts,
     configHash,
     cacheFullDir,
   };
